refactor(ble): turn the 556-line createNobleHandler closure into modules
src/ble/handler-noble-shared.ts held seven nested functions and their helpers
in one factory closure, making it the largest single construct in the project.
It is now src/ble/handler-noble-shared/, split along the banner comments that
already marked the seams.

Nine modules rather than the five the issue sketched, because a third of the
file was never in that table: wrapChar and the CCCD fallback, the peripheral
address and manufacturer-data helpers, the GATT wrappers, and the two exported
types. Leaving them behind would have made index.ts the same file with fewer
functions in it.

Three things the plan got wrong and the code corrected:

  - only waitForPoweredOn uses getState. connectWithRetries, discoverPeripheral
    and broadcastScan close over the Noble instance alone, so they take it as
    one parameter rather than a context object.

  - there is no shared mutable state between the four. adapterWarningLogged is
    the only mutable value in the factory, it belongs to scanAndReadRaw, and it
    stays inside the closure: hoisting it to module scope would make it
    process-global across handler instances, which the tests construct five of.

  - _internals.broadcastScan cannot point at the import, because the extracted
    function takes the Noble instance first. It is bound here so the six tests
    that call it with three arguments keep working unchanged.

Verified as a pure move: comparing the old file against the concatenation of
the new modules, ignoring imports, every difference is a signature gaining
`export` or the noble parameter, the matching call sites, or the _internals
binding. No logic line moved, changed or disappeared.

Both driver entrypoints keep createNobleHandler({ noble, getState }) and change
only their import path. The full suite passes with no assertion edits.

index.ts is 320 lines, above the ~200 the issue asks for. Splitting scanAndReadRaw
further is the same problem as #368 and belongs with it.

Closes #369
